requirements.txt
Pillow
numpy
exifread
